added sweet alert on delete button

+++ src/Pages/AllToys/AllToys.jsx
@@ -2,6 +2,7 @@ import { useLoaderData } from "react-router-dom";
 import useTitle from "../../Hooks/useTitle";
 import SingleToyRow from "./SingleToyRow";
 import { useState } from "react";
+import Swal from "sweetalert2";
 
 const AllToys = () => {
   useTitle("All toys");
@@ -9,19 +10,32 @@ const AllToys = () => {
   const [allToys, setAllToys] = useState(allUserData);
 
   const handleDeleteToy = (id) => {
-    fetch(`http://localhost:5000/allToys/${id}`, {
-      method: "DELETE",
-      headers: {
-        "content-type": "application/json",
-      },
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          const remaining = allToys.filter((toy) => toy._id !== id);
-          setAllToys(remaining);
-        }
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/allToys/${id}`, {
+          method: "DELETE",
+          headers: {
+            "content-type": "application/json",
+          },
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+              const remaining = allToys.filter((toy) => toy._id !== id);
+              setAllToys(remaining);
+            }
+          });
+      }
+    });
   };
 
   return (
